Consolidate duplicate tests
