+++ test/generator/entryClass.test.js
@@ -1,8 +1,9 @@
 import { test, expect } from '@jest/globals';
 import { generateBlogOuter } from '../../src/generator/generator.js';
 
-test('generateBlogOuter includes entry class in article and sections', () => {
-  const blog = { posts: [{ key: 'a1', content: ['hello'] }] };
+test('generateBlogOuter applies entry class to header, article, and footer', () => {
+  const blog = { posts: [{ key: 'a1', title: 't', publicationDate: '2024-01-01', content: ['hello'] }] };
   const html = generateBlogOuter(blog);
-  expect(html).toContain('class="entry"');
+  const matches = html.match(/class="entry"/g) || [];
+  expect(matches.length).toBe(3);
 });
