--- test/generator/footerWarningClass.count.test.js
@@ -1,8 +1,0 @@
-import { test, expect } from '@jest/globals';
-import { generateBlogOuter } from '../../src/generator/generator.js';
-
-test('footer warning class appears exactly once', () => {
-  const html = generateBlogOuter({ posts: [] });
-  const matches = html.match(/<div class="footer value warning">/g) || [];
-  expect(matches.length).toBe(1);
-});
