+++ test/generator/footerWarningClass.test.js
@@ -1,10 +1,8 @@
-import { describe, it, expect } from '@jest/globals';
+import { test, expect } from '@jest/globals';
 import { generateBlogOuter } from '../../src/generator/generator.js';
 
-describe('footer warning class', () => {
-  it('includes the warning CSS class on the footer value div', () => {
-    const html = generateBlogOuter({ posts: [] });
-    const regex = /<div class="footer value warning">/;
-    expect(html).toMatch(regex);
-  });
+test('footer warning class appears exactly once', () => {
+  const html = generateBlogOuter({ posts: [] });
+  const matches = html.match(/<div class="footer value warning">/g) || [];
+  expect(matches.length).toBe(1);
 });
